+++ binance/spot/_bswap.py
