+++ binance/websocket/spot/websocket_api/__init__.py
@@ -28,43 +28,43 @@ class SpotWebsocketAPIClient(BinanceWebsocketClient):
         )
 
     # Market
-    from binance.websocket.spot.websocket_api.market import ping_connectivity
-    from binance.websocket.spot.websocket_api.market import server_time
-    from binance.websocket.spot.websocket_api.market import exchange_info
-    from binance.websocket.spot.websocket_api.market import order_book
-    from binance.websocket.spot.websocket_api.market import recent_trades
-    from binance.websocket.spot.websocket_api.market import historical_trades
-    from binance.websocket.spot.websocket_api.market import aggregate_trades
-    from binance.websocket.spot.websocket_api.market import klines
-    from binance.websocket.spot.websocket_api.market import ui_klines
-    from binance.websocket.spot.websocket_api.market import avg_price
-    from binance.websocket.spot.websocket_api.market import ticker_24hr
-    from binance.websocket.spot.websocket_api.market import ticker
-    from binance.websocket.spot.websocket_api.market import ticker_price
-    from binance.websocket.spot.websocket_api.market import ticker_book
+    from binance.websocket.spot.websocket_api._market import ping_connectivity
+    from binance.websocket.spot.websocket_api._market import server_time
+    from binance.websocket.spot.websocket_api._market import exchange_info
+    from binance.websocket.spot.websocket_api._market import order_book
+    from binance.websocket.spot.websocket_api._market import recent_trades
+    from binance.websocket.spot.websocket_api._market import historical_trades
+    from binance.websocket.spot.websocket_api._market import aggregate_trades
+    from binance.websocket.spot.websocket_api._market import klines
+    from binance.websocket.spot.websocket_api._market import ui_klines
+    from binance.websocket.spot.websocket_api._market import avg_price
+    from binance.websocket.spot.websocket_api._market import ticker_24hr
+    from binance.websocket.spot.websocket_api._market import ticker
+    from binance.websocket.spot.websocket_api._market import ticker_price
+    from binance.websocket.spot.websocket_api._market import ticker_book
 
     # Account
-    from binance.websocket.spot.websocket_api.account import account
-    from binance.websocket.spot.websocket_api.account import order_rate_limit
-    from binance.websocket.spot.websocket_api.account import order_history
-    from binance.websocket.spot.websocket_api.account import oco_history
-    from binance.websocket.spot.websocket_api.account import my_trades
-    from binance.websocket.spot.websocket_api.account import prevented_matches
+    from binance.websocket.spot.websocket_api._account import account
+    from binance.websocket.spot.websocket_api._account import order_rate_limit
+    from binance.websocket.spot.websocket_api._account import order_history
+    from binance.websocket.spot.websocket_api._account import oco_history
+    from binance.websocket.spot.websocket_api._account import my_trades
+    from binance.websocket.spot.websocket_api._account import prevented_matches
 
     # Trade
-    from binance.websocket.spot.websocket_api.trade import new_order
-    from binance.websocket.spot.websocket_api.trade import new_order_test
-    from binance.websocket.spot.websocket_api.trade import get_order
-    from binance.websocket.spot.websocket_api.trade import cancel_order
-    from binance.websocket.spot.websocket_api.trade import cancel_replace_order
-    from binance.websocket.spot.websocket_api.trade import get_open_orders
-    from binance.websocket.spot.websocket_api.trade import cancel_open_orders
-    from binance.websocket.spot.websocket_api.trade import new_oco_order
-    from binance.websocket.spot.websocket_api.trade import get_oco_order
-    from binance.websocket.spot.websocket_api.trade import cancel_oco_order
-    from binance.websocket.spot.websocket_api.trade import get_open_oco_orders
+    from binance.websocket.spot.websocket_api._trade import new_order
+    from binance.websocket.spot.websocket_api._trade import new_order_test
+    from binance.websocket.spot.websocket_api._trade import get_order
+    from binance.websocket.spot.websocket_api._trade import cancel_order
+    from binance.websocket.spot.websocket_api._trade import cancel_replace_order
+    from binance.websocket.spot.websocket_api._trade import get_open_orders
+    from binance.websocket.spot.websocket_api._trade import cancel_open_orders
+    from binance.websocket.spot.websocket_api._trade import new_oco_order
+    from binance.websocket.spot.websocket_api._trade import get_oco_order
+    from binance.websocket.spot.websocket_api._trade import cancel_oco_order
+    from binance.websocket.spot.websocket_api._trade import get_open_oco_orders
 
     # User Data Stream
-    from binance.websocket.spot.websocket_api.user_data import user_data_start
-    from binance.websocket.spot.websocket_api.user_data import user_data_ping
-    from binance.websocket.spot.websocket_api.user_data import user_data_stop
+    from binance.websocket.spot.websocket_api._user_data import user_data_start
+    from binance.websocket.spot.websocket_api._user_data import user_data_ping
+    from binance.websocket.spot.websocket_api._user_data import user_data_stop
